+++ include/ccf/ds/json_schema.h
@@ -64,6 +64,28 @@ namespace ccf::ds::json
     return element;
   }
 
+  /** Produces the schema for a field which is marked as required (ie -
+   * always present in the JSON object), but whose C++ type is
+   * std::optional<T>. Such fields are always serialised, but may hold a JSON
+   * null when the C++ value is std::nullopt (for instance, a consensus's
+   * primary_id while no primary is currently known). The produced schema
+   * therefore describes the inner type T, additionally marked as nullable.
+   */
+  template <typename T>
+  inline nlohmann::json required_schema_element()
+  {
+    if constexpr (ccf::nonstd::is_specialization<T, std::optional>::value)
+    {
+      auto element = schema_element<typename T::value_type>();
+      element["nullable"] = true;
+      return element;
+    }
+    else
+    {
+      return schema_element<T>();
+    }
+  }
+
   namespace adl
   {
 #pragma clang diagnostic push